rename example `App.tsx` -> `ExampleApp.tsx`
