+++ examples/excalidraw/components/ExampleApp.scss
